Make checkMapKey code equal to the one in master
The differences were probably due to a different redis module being
used whereas a "null" value was returned for a missing key in the
redis version in "master" and nothing is returned instead in "develop".

A "null" value, in master branch, was giving everyone access to
a database with _NO_ known map key registered.